# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для сервиса справочники",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/countries": {
            "get": {
                "tags": [
                    "countries"
                ],
                "summary": "List of countries",
                "description": "Get list of countries",
                "operationId": "getCountries",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
        "/api/country": {
            "post": {
                "tags": [
                    "countries"
                ],
                "summary": "Create country",
                "description": "Create country",
                "operationId": "createCountry",
                "requestBody": {
                    "description": "Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Kazakhstan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
        "/api/country/{id}": {
            "get": {
                "tags": [
                    "countries"
                ],
                "summary": "Get country",
                "description": "Get country",
                "operationId": "getCountry",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "country ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            },
            "put": {
                "tags": [
                    "countries"
                ],
                "summary": "Update country",
                "description": "Update country",
                "operationId": "updateCountry",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "country ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            },
            "delete": {
                "tags": [
                    "countries"
                ],
                "summary": "Delete country",
                "description": "Delete country",
                "operationId": "deleteCountry",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "country ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
...
    },
...
}